requirements.txt
# ─── Web framework ───────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=13.0

# ─── Config ───────────────────────────────────────────────────────────────────
pydantic-settings>=2.3.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# ─── Redis ────────────────────────────────────────────────────────────────────
redis[hiredis]>=5.0.6

# ─── HTTP client (internal service calls) ────────────────────────────────────
httpx>=0.27.0

# ─── MCP SDK ─────────────────────────────────────────────────────────────────
mcp>=1.0.0

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0

# ─── Rate limiting ────────────────────────────────────────────────────────────
slowapi>=0.1.9

# ─── Structured logging ───────────────────────────────────────────────────────
structlog>=24.1.0

# ─── E2B sandbox ─────────────────────────────────────────────────────────────
e2b>=0.17.0
e2b-code-interpreter>=0.0.10

# ─── Composio ────────────────────────────────────────────────────────────────
composio-core>=0.5.0

# ─── AWS S3 (tool file sync – preserved from v1) ─────────────────────────────
boto3>=1.34.0

# ─── AI / tool dependencies (used by local plugin tools) ─────────────────────
openai>=1.0.0
google-generativeai>=0.7.0
google-cloud-aiplatform>=1.50.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0

# ─── Common tool dependencies ────────────────────────────────────────────────
requests>=2.31.0
pillow>=10.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
fpdf2>=2.7.0
beautifulsoup4>=4.12.0
pytz>=2024.1

# ─── Dev / testing ────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0
